Relocating input files
