+++ src/TWidgets/Core/Input/IInputWidget.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using TWidgets.Core.Drawing;
-using TWidgets.Core.IO;
 
-namespace TWidgets.Core
+namespace TWidgets.Core.Input
 {
     public interface IInputWidget
     {
