+++ src/TWidgets/Core/Input/InputCursor.cs
@@ -1,4 +1,4 @@
-﻿namespace TWidgets.Core.IO
+﻿namespace TWidgets.Core.Input
 {
     public class InputCursor
     {
