+++ src/TWidgets/Core/Input/InputMethod.cs
@@ -1,4 +1,4 @@
-﻿namespace TWidgets.Core.IO
+﻿namespace TWidgets.Core.Input
 {
     public enum InputMethod
     {
