+++ src/TWidgets/Core/Input/ValidateAction.cs
@@ -1,4 +1,4 @@
-﻿namespace TWidgets.Core.IO
+﻿namespace TWidgets.Core.Input
 {
     public enum ValidateAction
     {
