+++ src/TWidgets/Core/Input/ValidationState.cs
@@ -1,4 +1,4 @@
-﻿namespace TWidgets.Core.IO
+﻿namespace TWidgets.Core.Input
 {
     public enum ValidationState
     {
